requirements.txt
pandas
fastapi
pyarrow
requests
sqlalchemy
python-decouple
psycopg2-binary
postgres
uvicorn
pydantic